fix: change 'id' type from usize into i32

+++ todo-backend/src/todo.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize)]
 struct ToDo {
-    id: usize,
+    id: i32,
     content: String,
     date: String,
     deadline: String,
@@ -42,7 +42,7 @@ pub async fn add(State(state): State<AppState>, Json(payload): Json<NewToDo>) ->
 // delete TODO
 // inputs: ID
 // output: "Delete the TODO" or "Error: the TODO isn't deleted"
-pub async fn delete(State(state): State<AppState>, Path(id): Path<usize>) -> &'static str {
+pub async fn delete(State(state): State<AppState>, Path(id): Path<i32>) -> &'static str {
     let result = sqlx::query!("DELETE FROM todo_list WHERE id = ?", id)
         .execute(&state.database_pool)
         .await;
